hugepage: add support to provision hugepages
add optional flow to support the provisioning of hugepages.
Initially, we only support dynamic provisioning at runtime,
which is the discouraged approach but is much simpler to
implement as it doesn't require a node reboot.

Later we will add support to generate the kernel commandline
to provision hugepages at node start, reusing the API
and the infrastructure added here.

Provisioning support makes sense to complete the functionality,
so this driver becomes capable to enumarate, allocate and provision
hugepages.

Online on-demand provisioning is intentionally not (yet?)
supported: all hugepages need to be provisioned before use,
this feature just makes it easier to do, reducing or avoiding
the need to use external tools/scripts.

Signed-off-by: Francesco Romani <fromani@redhat.com>