driver: sysinfo: rename package
prepare to generalize the machine info gathering, making room.
The `ghwinfo` package becomes `sysinfo`, and we now extract
the topology bits we care about (per-numa memory report) and add
the missing informations, starting with the memory page size.
Overall, this is not much more work but a cleaner more targeted
approach.

The memory page size will be needed later to correctly configure
the memory and hugetlb hard limits.

Signed-off-by: Francesco Romani <fromani@redhat.com>