alloc: add allocation tracker
Introducing the hugepages support means we now
need to track resources in the plugin and bridge between
DRA and NRI layers.

The on-disk CDI data still is a source of truth, but
the allocation tracker (/manager) is meant to simplify
the bookkeeping.

Later when we will introduce the (re-)synchronization support,
we will probably shift to the allocation tracker as source of truth.

Signed-off-by: Francesco Romani <fromani@redhat.com>