types: add common package
we can use a common set of types to represent
resources, which we will use across all the layers.
This will bring consistency and reduce duplication.

We use the new package in the `make-manifests` option,
fixing immediately the manifests generation and
bringing it on par with manually generated, added recently.

Signed-off-by: Francesco Romani <fromani@redhat.com>