attributes: move towards standard naming
Trying to keep the names proliferation in check, we
now move closer towards what seems to become the standard:
deviceattribute package prefixes. We also classify
our attributes in maturity classes, still trying to
simplify our attribute set.

The attribute naming is not final yet even though we
are slowly converging; a final decision needs to be
made before we cut the first release, even though
that release will almost surely be an alpha.

Signed-off-by: Francesco Romani <fromani@redhat.com>